Switches: fix basic slider decimals and tests
Set the decimals variant without calling QVariant::toInt(), otherwise it
is set to a valid QVariant(0) variant.

Fix the decimals/stepsize tests: actually check the expected output, add
additional step size tests, and fix the test input paths so that values
are set correctly.

Fixes #2866

+++ src/iochannel.cpp
@@ -258,7 +258,7 @@ int IOChannel::decimals() const
 
 void IOChannel::setDecimals(const QVariant &variant)
 {
-	m_decimalsVariant = variant.toInt();
+	m_decimalsVariant = variant;
 	updateDecimals();
 }
